+++ changelog.md
@@ -157,5 +157,6 @@ Now works with all dialects of SQL supported by Sequelize (MySQL, Postgres, SQLi
 
 ## Next
 
+* Run jshint on tests
 * Disable Travis dependency cache
 * Update README badges to use shields.io